:recycle: [#2685] -- refactor some translation utils for broader usage
Rather than duplicating reading out the settings.LANGUAGES,
we can re-use existing implementations to provide choices for
the django user model field.
